requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
tabulate>=0.9.0

# AI/LLM dependencies
google-genai>=1.0.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
statsmodels>=0.14.0

# Geocoding
geopy>=2.4.0

# Utilities
pillow>=10.0.0
